CryptoPrinter/requirements.txt
ccxt==4.1.13
openai==1.3.7
pandas==2.1.3
ta==0.10.2
python-dotenv==1.0.0
requests==2.31.0